Rename xidauth to ejabberd_xidauth.
We now have a general Python xidauth library.  The existing xidauth.py
script for ejabberd-specific authentification is renamed to ejabberd_xidauth
to avoid ambiguity.
