+++ ejabberd/docker/Dockerfile
@@ -1,5 +1,5 @@
 # This constructs a Docker image containing ejabberd (based mostly on
-# the official ejabberd/ecs image) as well as the xidauth.py extauth
+# the official ejabberd/ecs image) as well as the ejabberd_xidauth.py extauth
 # script and the required Python runtime for it.
 
 # To bypass https://github.com/moby/moby/issues/3465, we use docker-copyedit
@@ -18,7 +18,7 @@ RUN apk add --no-cache \
   py3-jsonrpclib
 
 WORKDIR $HOME
-COPY ejabberd/xidauth.py ejabberd/docker/xidauth.sh bin/
+COPY ejabberd/ejabberd_xidauth.py ejabberd/docker/ejabberd_xidauth.sh bin/
 
 USER ejabberd
 LABEL description="ejabberd/ecs image with XID authentication script"
