+++ ejabberd/ejabberd_xidauth.py
